Add a new action 'start_on_b_leg' to 'record_call' callflow
This fixes an issue where call recording files for both a-leg and b-leg were being saved
If the agent wanted to stop recording (to get the customers credit card details for example)
and then start recording again afterwards. A total of 3 call recording files were created:

1. recording of complete call on a-leg side started by by 'record_call' callflow 'start' action
2. recording between *7 start_record and *9 stop_record
3. recording after agent *7 start_record to restart recording and end of the call.

Obviously this is a big problem, the reason to stop recording was to NOT record the CC details but the details are still recorded in #1

A new 'record_call' action 'start_on_b_leg' now starts the call recording on the b-leg rather than the a-leg so in the same scenario only 2 call recording files are
created #2 and #3 in the list above. The CC details are no longer recorded.